Update Spatial Distribution of Hospitals in Fulton and DeKalb Counties .Rmd
The objective of this analysis is to investigate the spatial distribution of hospitals in Fulton and DeKalb counties, Georgia, and assess its relationship to socioeconomic factors to identify potential inequities in healthcare access across neighborhoods. 

This study aims to answer whether the current distribution of hospitals is equitable, considering demographic and economic disparities across Census Tracts.